+++ contracts/ModuleRegistry.sol
@@ -4,8 +4,8 @@ import "./interfaces/IModuleRegistry.sol";
 import "./interfaces/IModuleFactory.sol";
 import "./interfaces/ISecurityToken.sol";
 import "./interfaces/ISecurityTokenRegistry.sol";
+import "./interfaces/IRegistry.sol";
 import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
-import "./helpers/ReclaimFunds.sol";
 
 /**
 * @title ModuleRegistry
@@ -13,7 +13,7 @@ import "./helpers/ReclaimFunds.sol";
 * Anyone can register modules, but only those "approved" by Polymath will be allowed to everyone.
 */
 
-contract ModuleRegistry is IModuleRegistry, Ownable, ReclaimFunds {
+contract ModuleRegistry is IModuleRegistry, Ownable, IRegistry {
 
     // Mapping used to hold the type of module factory corresponds to the address of the Module factory contract
     mapping (address => uint8) public registry;
@@ -33,9 +33,6 @@ contract ModuleRegistry is IModuleRegistry, Ownable, ReclaimFunds {
     event LogModuleRegistered(address indexed _moduleFactory, address indexed _owner);
     // Emit when the module get verified by the Polymath team
     event LogModuleVerified(address indexed _moduleFactory, bool _verified);
-
-    // Payable Fallback
-    function() public payable {}
 
     /**
     * @dev Called by a security token to notify the registry it is using a module