+++ test/security_token_registry.js
@@ -408,6 +408,8 @@ contract('SecurityTokenRegistry', accounts => {
     describe("Generate SecurityToken v3", async() => {
 
         it("Should add the new custom token in the polymath network", async() => {
+            await I_PolyToken.getTokens((10000 * Math.pow(10, 18)), account_temp);
+            await I_PolyToken.approve(I_TickerRegistry.address, initRegFee, { from: account_temp});
             await I_TickerRegistry.registerTicker(account_temp, "CUST", "custom", "I am swram hash", {from: account_temp});
             let tx = await I_SecurityTokenRegistry.addCustomSecurityToken("custom", "CUST", account_temp, accounts[2], "I am custom ST", "I am swram hash", {from: account_polymath});
             assert.equal(tx.logs[0].args._symbol, "CUST");
@@ -565,29 +567,29 @@ contract('SecurityTokenRegistry', accounts => {
 
     });
 
-    describe("Test cases for setRegistrationFee", async() => {
+    describe("Test cases for the changePolyRegisterationFee", async() => {
 
         it("Should successfully get the registration fee", async() => {
             let fee = await I_SecurityTokenRegistry.registrationFee.call();
             assert.equal(fee, initRegFee)
         });
 
-        it("Should fail to set the registration fee if msg.sender not owner", async() => {
+        it("Should fail to change the registration fee if msg.sender not owner", async() => {
             let errorThrown = false;
             try {
-                let tx = await I_SecurityTokenRegistry.setPolyRegistrationFee(400 * Math.pow(10, 18), {from: account_temp});
+                let tx = await I_SecurityTokenRegistry.changePolyRegisterationFee(400 * Math.pow(10, 18), { from: account_temp });
             } catch(error) {
-                console.log(`         tx revert -> Failed to set registrationFee`.grey);
+                console.log(`         tx revert -> Failed to change registrationFee`.grey);
                 errorThrown = true;
                 ensureException(error);
             }
             assert.ok(errorThrown, message);
         });
 
-        it("Should successfully set the registration fee", async() => {
-            await I_SecurityTokenRegistry.setPolyRegistrationFee(400 * Math.pow(10, 18), {from: account_polymath});
+        it("Should successfully change the registration fee", async() => {
+            await I_SecurityTokenRegistry.changePolyRegisterationFee(400 * Math.pow(10, 18), { from: account_polymath });
             let fee = await I_SecurityTokenRegistry.registrationFee.call();
-            assert.equal(fee, 400 * Math.pow(10, 18))
+            assert.equal(fee, 400 * Math.pow(10, 18));
         });
 
     });
@@ -595,21 +597,11 @@ contract('SecurityTokenRegistry', accounts => {
     describe("Test cases for reclaiming funds", async() => {
 
         it("Should successfully reclaim POLY tokens", async() => {
-            I_PolyToken.transfer(I_TickerRegistry.address, 1 * Math.pow(10, 18), { from: token_owner });
+            I_PolyToken.transfer(I_SecurityTokenRegistry.address, 1 * Math.pow(10, 18), { from: token_owner });
             let bal1 = await I_PolyToken.balanceOf.call(account_polymath);
-            await I_TickerRegistry.reclaimERC20(I_PolyToken.address);
+            await I_SecurityTokenRegistry.reclaimERC20(I_PolyToken.address);
             let bal2 = await I_PolyToken.balanceOf.call(account_polymath);
             assert.isAbove(bal2, bal1);
-        });
-
-        it("Should successfully reclaim ETH", async() => {
-            await I_TickerRegistry.sendTransaction({ value: 1 * Math.pow(10, 18), from: account_polymath })
-            let bal1 = await web3.eth.getBalance(account_polymath);
-            await I_TickerRegistry.reclaimETH();
-            let bal2 = await web3.eth.getBalance(account_polymath);
-            let bal3 = await web3.eth.getBalance(I_TickerRegistry.address);
-            assert.isAbove(bal2, bal1);
-            assert.equal(bal3, 0);
         });
 
     });