+++ test/ticker_registry.js
@@ -318,30 +318,31 @@ contract('TickerRegistry', accounts => {
         });
     });
 
-    describe("Test cases for the setPolyRegistrationFee", async() => {
+    describe("Test cases for the changePolyRegisterationFee", async() => {
 
         it("Should successfully get the registration fee", async() => {
             let fee = await I_TickerRegistry.registrationFee.call();
             assert.equal(fee, initRegFee)
         });
 
-        it("Should fail to set the registration fee if msg.sender not owner", async() => {
+        it("Should fail to change the registration fee if msg.sender not owner", async() => {
             let errorThrown = false;
             try {
-                let tx = await I_TickerRegistry.setPolyRegistrationFee(400 * Math.pow(10, 18), {from: account_temp});
+                let tx = await I_TickerRegistry.changePolyRegisterationFee(400 * Math.pow(10, 18), { from: account_temp });
             } catch(error) {
-                console.log(`         tx revert -> Failed to set registrationFee`.grey);
+                console.log(`         tx revert -> Failed to change registrationFee`.grey);
                 errorThrown = true;
                 ensureException(error);
             }
             assert.ok(errorThrown, message);
         });
 
-        it("Should successfully set the registration fee", async() => {
-            await I_TickerRegistry.setPolyRegistrationFee(400 * Math.pow(10, 18), {from: account_polymath});
+        it("Should successfully change the registration fee", async() => {
+            await I_TickerRegistry.changePolyRegisterationFee(400 * Math.pow(10, 18), { from: account_polymath });
             let fee = await I_TickerRegistry.registrationFee.call();
             assert.equal(fee, 400 * Math.pow(10, 18));
         });
+
     });
 
     describe("Test cases for reclaiming funds", async() => {
@@ -352,16 +353,6 @@ contract('TickerRegistry', accounts => {
             await I_TickerRegistry.reclaimERC20(I_PolyToken.address);
             let bal2 = await I_PolyToken.balanceOf.call(account_polymath);
             assert.isAbove(bal2, bal1);
-        });
-
-        it("Should successfully reclaim ETH", async() => {
-            await I_TickerRegistry.sendTransaction({ value: 1 * Math.pow(10, 18), from: account_polymath })
-            let bal1 = await web3.eth.getBalance(account_polymath);
-            await I_TickerRegistry.reclaimETH();
-            let bal2 = await web3.eth.getBalance(account_polymath);
-            let bal3 = await web3.eth.getBalance(I_TickerRegistry.address);
-            assert.isAbove(bal2, bal1);
-            assert.equal(bal3, 0);
         });
 
     });